refactor(domain): move ports interfaces to domain
