+++ internal/eval/readers.go
@@ -1,15 +1,16 @@
 package eval
 
 import (
+	"github.com/b2wdigital/restQL-golang/internal/domain"
 	"github.com/pkg/errors"
 )
 
 type MappingsReader struct {
-	env  EnvSource
+	env  domain.EnvSource
 	file map[string]string
 }
 
-func NewMappingReader(config Configuration, log Logger) MappingsReader {
+func NewMappingReader(config domain.Configuration, log domain.Logger) MappingsReader {
 	mr := MappingsReader{env: config.Env()}
 
 	mappingsConf := struct {
@@ -43,7 +44,7 @@ type QueryReader struct {
 	file map[string]savedQueries
 }
 
-func NewQueryReader(config Configuration, log Logger) QueryReader {
+func NewQueryReader(config domain.Configuration, log domain.Logger) QueryReader {
 	queryConf := struct {
 		Queries map[string]savedQueries
 	}{}
