+++ internal/eval/runner/runner.go
@@ -1,0 +1,7 @@
+package runner
+
+import "github.com/b2wdigital/restQL-golang/internal/domain"
+
+type Runner struct {
+	config domain.Configuration
+}
